requirements.txt
# Voice AI Ordering Agent - Dependencies
# Install with: pip install -r requirements.txt

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
python-multipart==0.0.6

# Twilio Voice/SMS Integration
twilio==8.13.0

# AI/ML - Hugging Face Transformers
transformers==4.37.0
torch==2.1.2
accelerate==0.26.0
sentencepiece==0.1.99
tokenizers==0.15.0

# Audio Processing & Transcription (Whisper)
openai-whisper==20231117
librosa==0.10.1
soundfile==0.12.1
SpeechRecognition==3.10.1

# Vector Search (FAISS for RAG)
faiss-cpu==1.7.4
sentence-transformers==2.3.1

# Text-to-Speech
gTTS==2.5.0
pyttsx3==2.90

# Database
aiosqlite==0.19.0

# Optional: Redis for distributed sessions
redis==5.0.1

# Optional: OpenAI API (alternative LLM backend)
openai==1.10.0

# Optional: Stripe for real payments
stripe==7.12.0

# HTTP Client
aiohttp==3.9.3
requests==2.31.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Code Quality
black==24.1.1
isort==5.13.2
flake8==7.0.0

# Utilities
numpy==1.26.3
pydantic==2.5.3

